% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ahn3_deBilt}
\alias{ahn3_deBilt}
\title{AHN3 data for the Bilt}
\format{An object of class \code{RasterStack} of dimension 12500 x 10000 x 1.}
\usage{
ahn3_deBilt
}
\description{
data can be downloaded from \url{https://www.pdok.nl/nl/ahn3-downloads}
}
\section{Slots}{

\describe{
\item{\code{r_32cn1}}{kaartblad number for the bilt with height elevation data}
}}

\keyword{datasets}
