% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AWS.df}
\alias{AWS.df}
\title{AWS coordinates}
\format{A data frame with 46 rows and 9 variables:
\describe{
  \item{IT_DATETIME}{price, in US dollars}
  \item{DS_CODE}{weight of the diamond, in carats}
  \item{DS_NAME}{weight of the diamond, in carats}
  \item{DS_DESC}{weight of the diamond, in carats}
  \item{DS_LAT}{weight of the diamond, in carats}
  \item{DS_LON}{weight of the diamond, in carats}
  \item{DS_ALT}{weight of the diamond, in carats}
  \item{REH1.TG}{weight of the diamond, in carats}
  \item{REH1.Q_TG}{weight of the diamond, in carats}
}}
\usage{
AWS.df
}
\description{
Dataset contains lat lon of all the aws stations.
}
\examples{
data(AWS.df)
str(AWS.df)
summary(AWS.df)

}
\keyword{datasets}
