% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_guidelinesCriteria.R
\name{check_criteria}
\alias{check_criteria}
\title{Check and select WMO SC Guidelines}
\usage{
check_criteria(df, criteria_columnName)
}
\arguments{
\item{df}{dataframe to select the criteria}

\item{criteria_columnName}{String. Name of column name from df}
}
\value{
data frame with values for each Class for the shading, heat sources and vegetation height criteria
}
\description{
Check and select WMO SC Guidelines
}
\details{
Check and select appropriate WMO SC guidelines criteria
}
\author{
Jelle Stuurman
create_spatialPoint(X = , Y = , LAT LON = TRUE)
}
