% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_shp.R
\name{check_shpExists}
\alias{check_shpExists}
\title{Check and remove shapefile}
\usage{
check_shpExists(shape_path)
}
\arguments{
\item{shape_path}{path of shapefile}
}
\description{
Check if shapefile path exists and remove all assocaited files from memory disk
}
\details{
Check if Shp files and remove it
}
\author{
Jelle Stuurman
}
