% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_point.R
\name{create_spatialPoint}
\alias{create_spatialPoint}
\title{spatial point}
\usage{
create_spatialPoint(X, Y, LONLAT)
}
\arguments{
\item{X}{X coordidnate in RD New or WGS84 (LON)}

\item{Y}{Y coordidnate in RD New or WGS84 (LAT)}

\item{LONLAT}{Optional. Default FALSE. Set to TRUE if X and Y are in Longitude and Latitude format. Output will be in RD New format.}
}
\value{
sf spatial point in RD New coordinates format
}
\description{
create a sf spatial point in RD New coordinates
}
\details{
Create a sf spatial point from X/Y or LAT?LON coordinates
}
\author{
Jelle Stuurman
create_spatialPoint(X = , Y = , LAT LON = TRUE)
}
