% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffer.R
\name{createBuffer}
\alias{createBuffer}
\title{Create buffer}
\usage{
createBuffer(point, distance,
  crs = rgdal::CRSargs(CRS("+init=epsg:28992")))
}
\arguments{
\item{point}{Required. sf point in RD New format}

\item{distance}{Required. distance in metres}

\item{crs}{Optional. Default RD New projection string}
}
\value{
buffer
}
\description{
Create buffer in RD new format
}
\details{
Create buffer from point
}
\author{
Jelle Stuurman
}
