% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_classification.R
\name{determine_temperatureSC}
\alias{determine_temperatureSC}
\title{temperature SC}
\usage{
determine_temperatureSC(X = NULL, Y = NULL, name, sensor.name,
  AWS = FALSE, name.supplement = "", LONLAT = FALSE,
  criteria_columnName = "Criteria_Value",
  class_selection = "final_class", import_ahn = FALSE,
  import_bgt = FALSE, redownload_bgt = FALSE, redownload_ahn = FALSE,
  AHN3 = TRUE, solar_angles = FALSE,
  angle_selection_byIndexNr = "all", years = 2018, months = c(12,
  1:6), days = c(21), s_hour = 0, f_hour = 23,
  minutes.interval = 60, ahn_resolution = 0.5, ahn_radius = 500,
  bgt_radius = 150, include_shadow_angles = FALSE,
  calculate_shadow_angles = FALSE, read_only_shadow_values = FALSE,
  extract_method = "bilinear", radius = 100,
  full_circle_mask = FALSE, sensor_height = 0,
  vegetation_radius = 10, exportShp = FALSE, exportCSV = FALSE,
  printChart = FALSE, delete_ahn_sheets = TRUE,
  delete_bgt_gmls = TRUE, test = FALSE, summary_name.supplement = "")
}
\value{
\enumerate{
  \item summary: summary table with overall and per critera SC of each analysed location
  \item All: same as generate classifcation
}
}
\description{
Determine air temperature SC
}
\details{
temperatureSC
}
