% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesect_bgt.R
\name{find_land_use_and_classes}
\alias{find_land_use_and_classes}
\title{find BGT alnd use through intersect}
\usage{
find_land_use_and_classes(spatialpoint, name, addition = "", bgt.shape,
  land_use_criteria.df, class, go_to_next_class,
  criteria_columnName = "Criteria_Value", exportCSV = FALSE,
  exportShp = FALSE)
}
\arguments{
\item{spatialpoint}{single sf point in RD new coordinates}

\item{name}{of AWS or location}

\item{bgt.shape}{shapefile of BGT dataset in RD New format}
}
\value{
liat
\enumerate{
  \item df: data frame with all the results of found objects and relative surface areas within all the buffers and annulus
  \item buffers: vector with buffer distances that are ssociated to that Class criteria
  \item outer_buf: shapefile of outer buffer
  \item annulus: shapefile of annulus
  \item innerbuf: shapefiler of inner buffer
}
}
\description{
Fi9nd land use of BGt by intersecting different radius circular areas
}
\details{
intersect with BGT data
}
\author{
Jelle Stuurman
}
