% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_classification.R
\name{generate_classifications}
\alias{generate_classifications}
\title{single SC}
\usage{
generate_classifications(spatialpoint, name, sensor.name,
  name.supplement = "", AWS = FALSE, LONLAT = FALSE,
  criteria_columnName = "Criteria_Value",
  class_selection = "final_class", import_ahn = FALSE,
  import_bgt = FALSE, redownload_ahn = FALSE, redownload_bgt = FALSE,
  AHN3 = TRUE, solar_angles = FALSE,
  angle_selection_byIndexNr = "all", years, months, days, s_hour = 0,
  f_hour = 23, minutes.interval = 60, ahn_resolution = 0.5,
  ahn_radius = 500, bgt_radius = 150, include_shadow_angles = TRUE,
  calculate_shadow_angles = FALSE, radius = 100,
  read_only_shadow_values = FALSE, extract_method = " bilinear",
  full_circle_mask = FALSE, sensor_height = 0,
  vegetation_radius = 10, exportShp = FALSE, exportCSV = FALSE,
  printChart = FALSE, delete_ahn_sheets = TRUE,
  delete_bgt_gmls = TRUE)
}
\value{
\enumerate{
  \item summary:
  \item overview_shading_table:
  \item complete_shading_table:
  \item land_use_table:
  \item vegetation table:
  \item map:
}
}
\description{
generate single temperature SC
}
\details{
generate single temperature SC
}
