% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_name.R
\name{getName_trim}
\alias{getName_trim}
\title{Trim name}
\usage{
getName_trim(name, name.supplement = "")
}
\arguments{
\item{name}{Name of AWS}

\item{name.supplement}{OOptional. If an name.supplemental string needs to be added to AWS string name}
}
\value{
String with trimmed name
}
\description{
Trim name of AWS by removing spaces
}
\details{
Trim name of point by removing spaces
}
\author{
Jelle Stuurman
}
