% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_shadow_angle.R
\name{highest_shadow_angle}
\alias{highest_shadow_angle}
\title{highest shadow angle}
\usage{
highest_shadow_angle(spatialpoint, AWS = FALSE, x, shadow_angle_raw,
  minutes.interval)
}
\arguments{
\item{spatialpoint}{single sf point.}

\item{x}{single azimuth angle}

\item{shadow_angle_raw}{shadow angle}

\item{minutes.interval}{minutes interval}
}
\value{
highest shadow angle value
}
\description{
Correct highest possible shadow angle
}
\details{
Correct highest possible shadow angle
}
\author{
Jelle Stuurman
}
