% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ahn.R
\name{import_single_ahn}
\alias{import_single_ahn}
\title{import AHN}
\usage{
import_single_ahn(spatialpoint, name, name.supplement = "",
  resolution = 0.5, radius, raw_ahn = TRUE, terrain_ahn = TRUE,
  AHN3 = TRUE, delete_sheets = TRUE, redownload = FALSE)
}
\arguments{
\item{spatialpoint}{single sf point in RD new coordinates}

\item{name}{of AWS or location}

\item{resolution}{Default 0.5. Set resoution to 0.5 or 5}

\item{radius}{radius distance in metres}

\item{AHN3}{Default TRUE. Set to FALSE if AHN2 needs to be used.}
}
\value{
list
\enumerate{
  \item ahn: AHN2 or AHN3
  \item ahn_rasters: stack of raw/andor AHN rasters
}
}
\description{
download AHN2 or AHN3 dataset around a certain location to storage memory
}
\author{
Jelle Stuurman
}
