% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bgt.R
\name{import_single_bgt}
\alias{import_single_bgt}
\title{import BGT}
\usage{
import_single_bgt(spatialpoint, name, name.supplement = "",
  sensor.name = "", radius = 150, delete_raw_gmls = TRUE,
  redownload = FALSE)
}
\arguments{
\item{spatialpoint}{single sf point in RD new coordinates}

\item{name}{of AWS or location}

\item{radius}{Deafult 150. Set radius distance in metres}
}
\description{
downlaod BGT dataset to storage memory
}
\author{
Jelle Stuurman
}
