% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_rd.R
\name{map_rd}
\alias{map_rd}
\title{RD map}
\usage{
map_rd(spatialpoint, name, sensor.name, name.supplement = "",
  AWS = FALSE, buffers, vegetation_height_raster, vegetation_radius,
  AHN3 = TRUE, class = "", zoom = 17)
}
\arguments{
\item{spatialpoint}{single sf point in RD new coordinates}

\item{name}{of AWS or location}

\item{AHN3}{Default TRUE. Set to FALSE if AHN2 needs to be used.}
}
\value{
Rd map as leaflet object
}
\description{
create interactive RD map
}
\details{
RD map
}
\author{
Jelle Stuurman
}
