% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_raster.R
\name{mask_raster}
\alias{mask_raster}
\title{mask raster}
\usage{
mask_raster(spatialpoint, name, name.supplement = "", ahn_raster,
  AHN3 = TRUE, azimuth, radius)
}
\arguments{
\item{spatialpoint}{single sf point in RD new coordinates}

\item{name}{of AWS or location}

\item{ahn_raster}{AHN raster}

\item{AHN3}{Default TRUE. Set to FALSE if AHN2 needs to be used.}

\item{radius}{distance radius of raster in metres.}

\item{azimuith}{solar azimuth angle in degrees}
}
\value{
ahn mask
}
\description{
section mask of raster
}
\examples{


}
\author{
Jelle Stuurman
}
