% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_import_ahn.R
\name{multiple_import_ahn}
\alias{multiple_import_ahn}
\title{smultiple import ahn}
\usage{
multiple_import_ahn(X = NULL, Y = NULL, name_list, sensor.name = "",
  name.supplement = "", AWS = TRUE, LONLAT = FALSE, ahn_radius,
  resolution = 0.5, AHN3 = TRUE, raw_ahn = TRUE,
  terrain_ahn = TRUE, delete_sheets = TRUE, redownload_ahn = FALSE)
}
\arguments{
\item{name_list}{list with string names.}

\item{LONLAT}{Logic. Dfeault FALSE. Set to TRUE if spatialpoint is in WGS84}

\item{ahn_radius}{radius of AHN}

\item{AHN3}{Default TRUE. Set to FALSE if AHN2 needs to be used.}
}
\description{
multiple import ahn
}
\details{
multiple import AHN
}
\author{
Jelle Stuurman
}
