% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_import_bgt.R
\name{multiple_import_bgt}
\alias{multiple_import_bgt}
\title{smultiple import BGT}
\usage{
multiple_import_bgt(X = NULL, Y = NULL, name_list, AWS = FALSE,
  name.supplement = "", sensor.name = "", bgt_radius = 150,
  delete_raw_gmls = TRUE, redownload = TRUE)
}
\arguments{
\item{name_list}{list with string names.}

\item{bgt_radius}{radius of AHN}
}
\description{
multiple import BGT
}
\details{
multiple import BGT
}
\author{
Jelle Stuurman
}
