% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projected_shade_class.R
\name{projected_shade_classes}
\alias{projected_shade_classes}
\title{Projected shade classes}
\usage{
projected_shade_classes(name, data.path, AHN3 = TRUE,
  name.supplement = "", criteria_columnName = "Criteria_Value")
}
\arguments{
\item{name}{coordidnate in RD New or WGS84 (LON)}

\item{data.path}{data path of shadow angles csv}

\item{AHN3}{Default TRUE. Set to FALSE if AHN2 needs to be used.}
}
\value{
data frame with all Projected shading Cass of all calculated shadow angles
}
\description{
Projected shade classes
}
\details{
Projected shade classes
}
\author{
Jelle Stuurman
}
