% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_single_aws.R
\name{select_single_aws}
\alias{select_single_aws}
\title{Select AWS point}
\usage{
select_single_aws(aws.df = AWS.df, name, sensor.name = "site")
}
\arguments{
\item{name}{Required. Name of AWS}

\item{sensor.name}{Optional. Name of sensor name.}
}
\value{
sf spatial point in RD New coordinates format
}
\description{
Select coordinates of AWs based on provided name of AWS
}
\details{
Select AWS point
}
\author{
Jelle Stuurman
}
