% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_mask_raster.R
\name{simple_mask_raster}
\alias{simple_mask_raster}
\title{simple mask raster}
\usage{
simple_mask_raster(spatialpoint, name, name.supplement = "", ahn_raster,
  AHN3 = TRUE, radius)
}
\arguments{
\item{spatialpoint}{single sf point in RD new coordinates}

\item{name}{of AWS or location}

\item{ahn_raster}{AHN raster}

\item{AHN3}{Default TRUE. Set to FALSE if AHN2 needs to be used.}

\item{radius}{radius distance of raster in metres}
}
\value{
ahn mask
}
\description{
full circle mask of raster
}
\author{
Jelle Stuurman
}
