% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_angles.R
\name{solar_angles}
\alias{solar_angles}
\title{solar angles}
\usage{
solar_angles(lat = 54.32556, lon = 2.935833, elv = 42.7, jd = 3)
}
\arguments{
\item{lat}{latitude of the station in WGS84}

\item{lon}{longitude}

\item{elv}{elevation}

\item{jd}{julian day}
}
\value{
dataframe with variables lat (latitude), lon (longitude), elv (elevation), jd (julian day), azimuth, zenith,...
}
\description{
iets over wat de functie doet
}
\details{
Calculating the solar angles for AWS locations
}
\examples{

data(AWS.df)
solar_angles(jd=seq(1,20,by=0.1)) #for the first 20 days with an accuracy of approximately 2.5 hours
}
\author{
Jelle Stuurman
}
