% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{sun_shade_angles_chart}
\alias{sun_shade_angles_chart}
\title{sun shadow chart}
\usage{
sun_shade_angles_chart(name, data.path, AHN3 = FALSE,
  name.supplement = "", angle_selection_byIndexNr = "all",
  extract_method = "bilinear")
}
\arguments{
\item{name}{coordidnate in RD New or WGS84 (LON)}

\item{data.path}{data path of shadow angles csv}

\item{AHN3}{Default TRUE. Set to FALSE if AHN2 needs to be used.}
}
\value{
data frame with all Projected shading Cass of all calculated shadow angles
}
\description{
create sun shadow chart using ggplot2
}
\details{
sun shadow chart
}
\author{
Jelle Stuurman
}
