% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation_height_classes.R
\name{vegetation_classes}
\alias{vegetation_classes}
\title{vegetation height Classes}
\usage{
vegetation_classes(df, criteria_columnName = "Criteria_Value", name,
  name.supplement = "", AHN3 = FALSE, exportCSV = FALSE)
}
\arguments{
\item{df}{summary statistics dataframe. Must include the column Median.}

\item{AHN3}{Default TRUE. Set to FALSE if AHN2 needs to be used.}

\item{spatialpoint}{single sf point in RD new coordinates}
}
\value{
dataframe with Class value for the vegetation criteria
\enumerate{
  \item height_raster: Height raster determined by subtracting the terrain from the raw AHN raster
  \item df: summary statistics of results
  \item raw_mask: mask raster of raw AHN
  \item terrain_mask: mask raster of terrain AHN
}
}
\description{
Determine vegetation height classes
}
\details{
vegetation height classes
}
\author{
Jelle Stuurman
}
