% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation_height.R
\name{vegetation_height}
\alias{vegetation_height}
\title{vegetation height}
\usage{
vegetation_height(spatialpoint, name, name.supplement = "", radius,
  AHN3 = TRUE, exportCSV = FALSE)
}
\arguments{
\item{spatialpoint}{single sf point in RD new coordinates}

\item{name}{of AWS or location}

\item{radius}{radius distance in metres}

\item{AHN3}{Default TRUE. Set to FALSE if AHN2 needs to be used.}
}
\value{
list
\enumerate{
  \item height_raster: Height raster determined by subtracting the terrain from the raw AHN raster
  \item df: summary statistics of results
  \item raw_mask: mask raster of raw AHN
  \item terrain_mask: mask raster of terrain AHN
}
}
\description{
Determine vegetation height
}
\details{
vegetation height
}
\author{
Jelle Stuurman
}
