% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AWS.sp}
\alias{AWS.sp}
\title{spatial data}
\format{An object of class \code{SpatialPointsDataFrame} with 46 rows and 7 columns.}
\usage{
AWS.sp
}
\description{
some desc
}
\section{Slots}{

\describe{
\item{\code{DS_CODE}}{iets}
}}

\keyword{datasets}
