% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_angles.R
\name{shadow_angles}
\alias{shadow_angles}
\title{shadow angles}
\usage{
shadow_angles(spatialpoint, name, name.supplement = "", LONLAT = FALSE,
  ahn_mask, angle, maxDist, AHN3 = TRUE, read_only = FALSE,
  extract_method = "bilinear")
}
\arguments{
\item{spatialpoint}{point of AWS}

\item{name}{of AWS or location}

\item{LONLAT}{Optional. Default FALSE}

\item{ahn_mask}{raster of AHN.}

\item{angle}{in degrees, starting with 0 in the north over east}

\item{maxDist}{maximum Distance}

\item{AHN3}{Default TRUE. Set to FALSE if AHN2 needs to be used}
}
\value{
data frame with shadow angle
}
\description{
calculates the angles of to the obstacles for a point
}
\details{
Shadow angles from a height grid around a point

something
}
