% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_angles.R
\name{shadow_angles}
\alias{shadow_angles}
\title{shadow angles}
\usage{
shadow_angles(sp, grid)
}
\description{
calculates the angles of to the obstacles for a point
}
\examples{
#kaartblad AHN3 for the bilt equals 32cn1
#downloadlink: 
}
