% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_angles.R
\name{solar_angles}
\alias{solar_angles}
\title{solar angles}
\usage{
solar_angles(spatialpoint, AWS = FALSE, day, month, year, s_hour, f_hour,
  minutes.interval = 60)
}
\arguments{
\item{spatialpoint}{single sf point.}

\item{day}{vector list. Days of the month to determine solar angles.}

\item{month}{numerical vector list. Numerical months of the year to determine solar angles.}

\item{year}{vector list. years to determine solar angles}

\item{s_hour}{Numeric. start hour to determine solar angles. dEFULT 0}

\item{f_hour}{Numeric. finish hour to determine solar angles. Default 23.}

\item{minutes.interval}{default 60. numerical minutes interval}

\item{LONLAT}{Logic. Dfeault FALSE. Set to TRUE if spatialpoint is in WGS84}
}
\value{
dataframe with variables lat (latitude), lon (longitude), elv (elevation), jd (julian day), azimuth, zenith,... This includes all
}
\description{
Determine solarangles at the selected point for the determined year(s), moth(s) day(s) for the selected time interval.
}
\details{
Calculating the solar angles at locations
}
\author{
Jelle Stuurman
}
