% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_angles.R
\name{solar_angles}
\alias{solar_angles}
\title{Calculating the solar angles for AWS locations}
\usage{
solar_angles(lat, lon, elv, jd)
}
\arguments{
\item{lat}{latitude of the station in WGS84}

\item{lon}{longitude}

\item{elv}{elevation}

\item{jd}{julian day}
}
\description{
Calculating the solar angles for AWS locations
}
\examples{

data(AWS.df)
#solar_angles()

}
